# backend/src/warehouses/warehouses.controller.ts
import {
  Controller,
  Get,
  Post,
  Body,
  Patch,
  Param,
  Delete,
  UseGuards,
  Query,
} from '@nestjs/common';
import { WarehousesService } from './warehouses.service';
import { CreateWarehouseDto } from './dto/create-warehouse.dto';
import { UpdateWarehouseDto } from './dto/update-warehouse.dto';
import { JwtAuthGuard } from '../auth/guards/jwt-auth.guard';
import { GetUser } from '../common/decorators/get-user.decorator';

@Controller('warehouses')
@UseGuards(JwtAuthGuard)
export class WarehousesController {
  constructor(private readonly warehousesService: WarehousesService) {}

  @Post()
  create(@Body() createWarehouseDto: CreateWarehouseDto, @GetUser() user: any) {
    return this.warehousesService.create(createWarehouseDto, user.companyId);
  }

  @Get()
  findAll(@GetUser() user: any, @Query('page') page?: string, @Query('limit') limit?: string) {
    const pageNum = page ? parseInt(page, 10) : 1;
    const limitNum = limit ? parseInt(limit, 10) : 10;
    return this.warehousesService.findAll(user.companyId, pageNum, limitNum);
  }

  @Get(':id')
  findOne(@Param('id') id: string, @GetUser() user: any) {
    return this.warehousesService.findOne(id, user.companyId);
  }

  @Patch(':id')
  update(
    @Param('id') id: string,
    @Body() updateWarehouseDto: UpdateWarehouseDto,
    @GetUser() user: any,
  ) {
    return this.warehousesService.update(id, updateWarehouseDto, user.companyId);
  }

  @Delete(':id')
  remove(@Param('id') id: string, @GetUser() user: any) {
    return this.warehousesService.remove(id, user.companyId);
  }

  @Get(':id/inventory')
  getInventory(@Param('id') id: string, @GetUser() user: any) {
    return this.warehousesService.getInventory(id, user.companyId);
  }
}